[FEATURE] Added OnClickListener to ToolTip
